fix: cast amount to i64 for ulyn-pay compatibility
ulyn-pay expects amount as i64 but CheckoutPayload deserializes it as
f64 from the HTMX form. serde_json serializes 15000.0 as a float,
which fails deserialization to i64 on the ulyn-pay side, returning a
400/422 and causing 'Gagal membuat link pembayaran'.

+++ src/handlers.rs
@@ -514,7 +514,7 @@ pub async fn checkout(
 
         let body = serde_json::json!({
             "external_user_id": u.id.to_string(),
-            "amount": payload.amount,
+            "amount": payload.amount as i64,
             "customer_name": customer_name,
             "customer_email": customer_email,
             "metadata": { "tier": payload.tier, "lang": lang },
